Time: 1218 ms (42.08%), Space: 8 MB (5.25%) - LeetHub

+++ 1523-count-odd-numbers-in-an-interval-range/1523-count-odd-numbers-in-an-interval-range.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
-    int countOdds(int low, int high) {
+    long long countOdds(long long low, long long high) {
         long long count=0;
-        for(int i=low;i<=high;i++){
+        for(long long i=low;i<=high;i++){
             if(i%2!=0){
                 count++;
             }
